requirements.txt
asgiref==3.4.1
boto3==1.21.11
botocore==1.24.11
dj-database-url==0.5.0
Django==3.2
django-allauth==0.41.0
django-countries==7.2.1
django-crispy-forms==1.14.0
django-storages==1.12.3
gunicorn==20.1.0
jmespath==0.10.0
oauthlib==3.1.1
Pillow==9.0.0
psycopg2-binary==2.9.3
python3-openid==3.2.0
pytz==2021.3
requests-oauthlib==1.3.0
s3transfer==0.5.2
sqlparse==0.4.2
stripe==2.66.0
